plan-utils: create and move helper functions
